todo_backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==23.1.0
boto3==1.24.83
botocore==1.27.83
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
cleo==0.7.6
click==8.1.3
clikit==0.4.3
coverage==6.5.0
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==0.5.0
Django==3.2.12
django-background-tasks @ git+https://github.com/mindsightbr/django-background-tasks.git@7725c8231b6281c182567b8a5477953cb565e666
django-colorfield==0.7.2
django-compat==1.0.15
django-constance==2.9.1
django-cors-headers==3.13.0
django-db-geventpool==4.0.1
django-debug-toolbar==3.7.0
django-environ==0.8.1
django-extensions==3.2.1
django-filter==21.1
django-import-export==2.8.0
django-mailer==2.2
django-otp==1.1.3
django-parler==2.3
django-picklefield==3.1
django-storages==1.13.1
django-tenants==3.4.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-flex-fields==0.9.9
elastic-apm==6.12.0
et-xmlfile==1.1.0
filelock==3.8.0
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
honcho==1.1.0
html5lib==1.1
identify==2.5.5
idna==3.4
importlib-metadata==4.11.1
iniconfig==1.1.1
isodate==0.6.1
jeepney==0.8.0
jmespath==0.10.0
jsonschema==3.2.0
keyring==20.0.1
lockfile==0.12.2
lxml==4.6.5
MarkupPy==1.14
mindsight-auth-integration @ git+https://gitlab.com/mindsight/mindsight-auth-integration.git@83110fa15b1844c36a618d5138b62160312a051b
-e git+ssh://git@gitlab.com/mindsight/mindsight-powerplug.git@3574026d4f8f7bf48305adba0ae964fdb3fe2860#egg=mindsight_powerplug
model-bakery==1.7.0
msgpack==1.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.0
pastel==0.2.1
pathspec==0.11.1
pexpect==4.8.0
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
psycogreen==1.0.2
psycopg2==2.9.3
psycopg2-binary==2.9.4
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.14.11
pytest==7.1.3
pytest-django==4.5.2
pytest-mock==3.9.0
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
python3-saml==1.14.0
pytz==2021.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.8.0
s3transfer==0.6.0
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.5.11
typing_extensions==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
webencodings==0.5.1
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
xmlsec==1.3.13
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
